refactor(app): extract composer integration
