+++ src/app.rs
@@ -19,6 +19,7 @@ pub mod community_hierarchy;
 pub mod composer;
 pub mod composer_input_mapping;
 pub mod composer_input_paste;
+pub mod composer_integration;
 pub mod contact_avatars;
 pub mod download_worker;
 pub mod events;
